% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_polygons.R
\name{extr_polygons}
\alias{extr_polygons}
\title{Extract Landcover Classes as Multipolygons from Classified Image}
\usage{
extr_polygons(
  class_img,
  data_format = ".gpkg",
  class_col = "class",
  append = FALSE,
  out_dir = "output_polygons"
)
}
\arguments{
\item{class_img}{A raster object representing the classified image.}

\item{data_format}{The format of the output data files. Can be ".shp" for shapefiles or ".gpkg" for GeoPackages. Defaults to ".gpkg".}

\item{class_col}{The name of the column in \code{class_img} containing the class labels. Defaults to "class".}

\item{append}{Logical indicating whether to append to existing data files if they already exist. Defaults to FALSE.}

\item{out_dir}{The directory where the output files will be saved. Defaults to "output_polygons".}
}
\value{
A list of multipolygon sf objects, each representing a landcover class extracted from the classified image.
}
\description{
Extracts landcover classes from a classified image as multipolygons and saves them as shapefiles or GeoPackages.
}
\details{
This function extracts landcover classes from a classified image represented by a raster object. It loops over unique class labels in the specified column (\code{class_col}) of the raster and converts each class into a binary raster. Then, it converts the binary raster into polygons using the \code{as.polygons} function from the \code{sp} package. The polygons are further converted into sf objects using \code{st_as_sf} from the \code{sf} package and saved as shapefiles or GeoPackages in the specified output directory (\code{out_dir}). The function returns a list of multipolygon sf objects, with each object representing a landcover class.
}
\examples{
# Load required packages
library(sf)
library(terra)

# Create a classified image raster
class_img <- rast(matrix(sample(1:3, 100, replace = TRUE), nrow = 10))

# Extract polygons
extr_polygons(class_img)

}
