% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_features.R
\name{pp_features}
\alias{pp_features}
\title{Pre-process Training Features for Supervised Classification}
\usage{
pp_features(trainFeat, raster)
}
\arguments{
\item{trainFeat}{Training features, either an sf object or an object that can be converted to an sf object.}

\item{raster}{A raster object representing the spatial extent and coordinate reference system (CRS) to which the training features will be aligned.}
}
\value{
A pre-processed sf object of training features aligned with coordinate reference system (CRS) of the input raster.
}
\description{
Pre-processes training features to prepare them for a supervised classification using the "superClass" command from the RStoolbox package.
The function transforms the training features to a sf object if it wasn't already and reprojects it to the coordinate reference system (CRS) of the raster input.
}
\details{
This function pre-processes the training features to ensure they are in the same coordinate reference system (CRS) as the input raster. If the training features are not already in the sf format, they are converted to an sf object using \code{sf::st_as_sf()}. Then, if the CRS of the training features differs from the CRS of the raster, the function transforms the CRS of the training features to match the CRS of the raster using \code{sf::st_transform()}.
}
\examples{
# Load required packages
library(RStoolbox)
library(sf)

# Create a raster object
raster <- raster::raster(matrix(runif(100), nrow = 10))

# Create training features
trainFeat <- data.frame(x = runif(10), y = runif(10))
trainFeat <- sf::st_as_sf(trainFeat, coords = c("x", "y"))

# Pre-process training features
pp_features(trainFeat, raster)

}
