% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_bands.R
\name{rename_bands}
\alias{rename_bands}
\title{Rename Bands in a Raster}
\usage{
rename_bands(raster, sensor, subsetting = TRUE)
}
\arguments{
\item{raster}{A raster object containing bands to be renamed.}

\item{sensor}{A character string specifying the sensor type (e.g., "Landsat8", "Sentinel2").}

\item{subsetting}{Logical indicating whether to subset the raster before renaming bands. Defaults to TRUE.}
}
\value{
A raster object with renamed bands.
}
\description{
Renames bands in a raster object according to the specified sensor. The function identifies original band positions for the bands
Blue, Green, Red, NIR, SWIR1, SWIR2 according to the sensor and names them "Blue", "Green", "Red", "NIR", "SWIR1", "SWIR2".
}
\details{
This function renames bands in a raster object based on the sensor type. Supported sensors and corresponding bands:
\itemize{
\item Landsat 8: Blue, Green, Red, NIR, SWIR1, SWIR2
\item Landsat 7: Blue, Green, Red, NIR, SWIR1, SWIR2
\item Landsat 5: Blue, Green, Red, NIR, SWIR1, SWIR2
\item Sentinel 2: Blue, Green, Red, NIR, SWIR1, SWIR2
}

If \code{subsetting} is set to FALSE, band names are directly replaced without subsetting the raster.

If \code{subsetting} is TRUE (default), the function subsets the raster to include only the renamed bands.
}
\examples{
# read sample file of the RStoolboxExtensions package
Sebangau15 <- system.file("extdata", "Sebangau15.tif", package = "RStoolboxExtensions")
Sebangau15 <- rast_sample_read(Sebangau15)

names(Sebangau15)

Sebangau15 <- rename_bands(Sebangau15, sensor = "Landsat8", subsetting = TRUE)

names(Sebangau15)

}
