% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameBands.R
\name{renameBands}
\alias{renameBands}
\title{Renaming Band Layers of a Raster}
\usage{
renameBands(raster, sensor, subsetting = TRUE)
}
\arguments{
\item{raster}{A raster object. The band names of this raster will be renamed based on their
original column position. If the band positions were altered the naming will be incorrect.}

\item{sensor}{The sensor your raster scene was generated with. So far Landsat 5-8 and Sentinel2
sensors are supported.}

\item{subsetting}{a logical evaluating to TRUE or FALSE indicating whether the raster scene should
be subsetted to the renamed bands ("blue", "green", "red", "nir", "swir1", "swir2")}
}
\value{
Returns the input raster with the renamed band layers. If subsetting = TRUE all other
bands than the renamed ones are removed.
}
\description{
Function to rename (and subset) the band layers
"blue", "green", "red", "nir", "swir1" and "swir2"
}
\examples{
\dontrun{
your_raster <- terra::rast('path_to_your_raster')
your_raster_renamed <- rename_bands_fun(raster = your_raster, sensor = "Landsat8", subsetting = TRUE)
}


}
