% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_sample_read.R
\name{sf_sample_read}
\alias{sf_sample_read}
\title{Read Sample SF Data for RStoolboxExtensions Package}
\usage{
sf_sample_read(path)
}
\arguments{
\item{path}{path to filename}
}
\value{
a \code{sf} that can be used as test data for the RStoolboxExtensions package functions.

#' @details included files are:
trainPoints.geojson - training points for landcover change detection of the Sebangau Nationalpark, Borneo between 2015-2023
trainPolygons.geojson - training polygons for landcover change detection of the Sebangau Nationalpark, Borneo between 2015-2023
}
\description{
This function can be used to load the sample sf data included in the RStoolboxExtensions package
into your environment. It serves as a demonstration how to retrieve the sf data imbedded in the RStoolboxExtensions package.
}
\examples{
\dontrun{
# retrieving the file where the sf object is stored and load it
trainPoints <- system.file("extdata", "trainPoints.geojson", package = "RStoolboxExtensions")
trainPoints <- sf_sample_read(trainPoints)

trainPolygons <- system.file("extdata", "trainPolygons.geojson", package = "RStoolboxExtensions")
trainPolygons <- sf_sample_read(trainPolygons)
}
}
