% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices.R
\name{calc_indices}
\alias{calc_indices}
\title{Calculate Spectral Indices}
\usage{
calc_indices(raster, indices = c("ndvi", "ndwi", "ndbi", "ndmi"))
}
\arguments{
\item{raster}{A raster object containing bands for Green, Red, NIR and SWIR1. The naming has to be accurate.
(e.g. the red band has to be named "Red", etc.)}

\item{indices}{A character vector specifying which indices to calculate. Defaults to c("ndvi", "ndwi", "ndbi", "ndmi").}
}
\value{
A raster object with calculated spectral indices added as additional layers.
}
\description{
Calculates specified spectral indices from multispectral raster data.
Supported indices include NDVI, NDWI, NDBI, and NDMI.
}
\details{
This function calculates the specified indices using the following formulas:
\deqn{NDVI = \frac{NIR - Red}{NIR + Red}}
\deqn{NDWI = \frac{Green - NIR}{Green + NIR}}
\deqn{NDBI = \frac{SWIR1 - NIR}{SWIR1 + NIR}}
\deqn{NDMI = \frac{NIR - SWIR1}{NIR + SWIR1}}
}
\examples{
# running the example may take up to a few minutes as raster processing takes its time.
# I recommend to run the example in a script to get update messages in the console.

# read sample file of the RStoolboxExtensions package
Sebangau15 <- system.file("extdata", "Sebangau15.tif", package = "RStoolboxExtensions")
Sebangau15 <- rast_sample_read(Sebangau15)

# rename the bands to have the correct band names for indices calculation
Sebangau15 <- rename_bands(Sebangau15, sensor = "Landsat8", subsetting = TRUE)

# calculate the indices of your choice
Sebangau15_indices <- calc_indices(Sebangau15, indices = c("ndvi", "ndbi", "ndmi"))

names(Sebangau15_indices)
head(Sebangau15_indices)

}
