% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices.R
\name{calc_indices}
\alias{calc_indices}
\title{Calculate Spectral Indices}
\usage{
calc_indices(raster, indices = c("ndvi", "ndwi", "ndbi", "ndmi"))
}
\arguments{
\item{raster}{A raster object containing bands for Green, Red, NIR and SWIR1. The naming has to be accurate.
(e.g. the red band has to be named "Red", etc.)}

\item{indices}{A character vector specifying which indices to calculate. Defaults to c("ndvi", "ndwi", "ndbi", "ndmi").}
}
\value{
A raster object with calculated spectral indices added as additional layers.
}
\description{
Calculates specified spectral indices from multispectral raster data.
Supported indices include NDVI, NDWI, NDBI, and NDMI.
}
\details{
This function calculates the specified indices using the following formulas:
\deqn{NDVI = \frac{NIR - Red}{NIR + Red}}
\deqn{NDWI = \frac{Green - NIR}{Green + NIR}}
\deqn{NDBI = \frac{SWIR1 - NIR}{SWIR1 + NIR}}
\deqn{NDMI = \frac{NIR - SWIR1}{NIR + SWIR1}}
}
\examples{
raster <- terra::rast(nir = rast(matrix(runif(100), nrow = 10)),
                red = rast(matrix(runif(100), nrow = 10)),
                green = rast(matrix(runif(100), nrow = 10)),
                swir1 = rast(matrix(runif(100), nrow = 10)))
calc_indices(raster)

}
