% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_rasters.R
\name{extr_rasters}
\alias{extr_rasters}
\title{Crops and Masks a Raster to a list of (Multi-)polygons}
\usage{
extr_rasters(
  raster,
  polygon_list,
  out_dir = "output_masked",
  overwrite = TRUE,
  datatype = "tif"
)
}
\arguments{
\item{raster}{A raster object to be cropped and masked.}

\item{polygon_list}{A list of (multi-)polygon sf objects, each representing a landcover class.}

\item{out_dir}{The directory where the output masked rasters will be saved. Defaults to "output_masked".}

\item{overwrite}{Logical indicating whether to overwrite existing files if they already exist. Defaults to TRUE.}

\item{datatype}{The data type of the output raster files. Defaults to "tif".}
}
\value{
A list of cropped and masked raster objects, each corresponding to an entry of the \code{polygon_list}.
}
\description{
Crops and masks a raster to a list of (multi-)polygons and saves the masked rasters as GeoTIFF or other raster formats.
}
\examples{
# loading sample multitpolygons from the RStoolboxExtensions package
afforestation <- system.file("extdata", "afforestation.gpkg", package = "RStoolboxExtensions")
afforestation <- sf_sample_read(afforestation)
deforestation <- system.file("extdata", "deforestation.gpkg", package = "RStoolboxExtensions")
deforestation <- sf_sample_read(deforestation)
nonforest <- system.file("extdata", "nonforest.gpkg", package = "RStoolboxExtensions")
nonforest <- sf_sample_read(nonforest)
forest <- system.file("extdata", "forest.gpkg", package = "RStoolboxExtensions")
forest <- sf_sample_read(forest)

# load classified sample image of the RStoolboxExtensions package
class_img <- system.file("extdata", "class_img.tif", package = "RStoolboxExtensions")
class_img <- rast_sample_read(class_img)

# put them together in a list
class_polygons <- list(c(afforestation, deforestation, nonforest, forest))

# apply the function
class_rasters <- extr_rasters(class_img, class_polygons, out_dir = "test_output_rasters")

View(class_rasters)

}
