% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_rasters.R
\name{extr_rasters}
\alias{extr_rasters}
\title{Crops and Mask Rasters to (Multi-)polygons}
\usage{
extr_rasters(
  raster,
  polygon_list,
  out_dir = "output_masked",
  overwrite = TRUE,
  datatype = "tif"
)
}
\arguments{
\item{raster}{A raster object to be cropped and masked.}

\item{polygon_list}{A list of (multi-)polygon sf objects, each representing a landcover class.}

\item{out_dir}{The directory where the output masked rasters will be saved. Defaults to "output_masked".}

\item{overwrite}{Logical indicating whether to overwrite existing files if they already exist. Defaults to TRUE.}

\item{datatype}{The data type of the output raster files. Defaults to "tif".}
}
\value{
A list of cropped and masked raster objects, each corresponding to an entry of the \code{polygon_list}.
}
\description{
Crops and masks a raster to a list of (multi-)polygons and saves the masked rasters as GeoTIFF or other raster formats.
}
\examples{
# Load required packages
library(sf)
library(terra)

# Create a raster object
raster <- rast(nrow=100, ncol=100)
values(raster) <- matrix(runif(10000), nrow=100)

# Create multipolygon sf objects representing landcover classes
polygons_list <- list(class1 = st_buffer(st_sfc(st_point(c(10, 10))), 10),
                      class2 = st_buffer(st_sfc(st_point(c(50, 50))), 10))

# Extract and mask rasters
extr_rasters(raster, polygons_list)

}
