% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_features.R
\name{pp_features}
\alias{pp_features}
\title{Pre-process Training Features for Supervised Classification}
\usage{
pp_features(trainFeat, raster)
}
\arguments{
\item{trainFeat}{Training features, either an sf object or an object that can be converted to an sf object.}

\item{raster}{A raster object representing the spatial extent and coordinate reference system (CRS) to which the training features will be aligned.}
}
\value{
A pre-processed sf object of training features aligned with coordinate reference system (CRS) of the input raster.
}
\description{
Pre-processes training features to prepare them for a supervised classification using the "superClass" command from the RStoolbox package.
The function transforms the training features to a sf object if it wasn't already and reprojects it to the coordinate reference system (CRS) of the raster input.
}
\details{
This function uses \code{sf::st_as_sf()} to convert to a sf object and \code{sf::st_transform()} to align the CRS.
}
\examples{
# read sample sf file of the RStoolboxExtensions package
trainPoints <- system.file("extdata", "trainPoints.geojson", package = "RStoolboxExtensions")
trainPoints <- sf_sample_read(trainPoints)

# read sample raster file of the RStoolboxExtensions package
Sebangau15 <- system.file("extdata", "Sebangau15.tif", package = "RStoolboxExtensions")
Sebangau15 <- rast_sample_read(Sebangau15)

# run the function
pp_trainPoints <- pp_features(trainPoints, Sebangau15)

# checking the class - should be sf
class(pp_trainPoints)

# checking the crs - should be aligned with raster crs
sf::st_crs(pp_trainPoints)
sf::st_crs(Sebangau15)


}
